Reorganize project structure and clean up code
Move source files into src/core/, src/ui/, src/games/ subdirectories.
Extract helpers (launch_game, handle_install_key, footer_key), add PATHBUF
constant, fix format_size signature, add default cases to switches.

+++ src/core/util.h
@@ -35,6 +35,6 @@ void  clear_install_method(const char *game_name);
 void  sanitize_name(const char *in, char *out, size_t outlen);
 
 /* ── Size formatting ──────────────────────────────────────────── */
-void  format_size(unsigned long bytes, char *buf, int buflen);
+void  format_size(long long bytes, char *buf, int buflen);
 
 #endif
